Combined hp and stat calculators
Single method to calculate a pokemon's new stats on level up. Core
formula is run, then extra arithmetic is added based on if the stat is
hp (unique formula), or one of the other stat types. A few variables to
not exist yet for this equation so it will not work.